* refactoring unit tests for clear block state.
