+++ crates/db/src/schema.rs
@@ -3,7 +3,7 @@
 /// Column families alias type
 pub type Col = &'static str;
 /// Total column number
-pub const COLUMNS: u32 = 18;
+pub const COLUMNS: u32 = 19;
 /// Column store meta data
 pub const COLUMN_META: Col = "0";
 /// Column store chain index
@@ -40,10 +40,8 @@ pub const COLUMN_DATA: Col = "15";
 pub const COLUMN_BLOCK_DEPOSITION_REQUESTS: Col = "16";
 /// Column custodian assets
 pub const COLUMN_CUSTODIAN_ASSETS: Col = "17";
-/// Column block account record of SMT branch
-pub const COLUMN_BLOCK_ACCOUNT_SMT_BRANCH_RECORD: Col = "18";
-/// Column block account record of SMT leaf
-pub const COLUMN_BLOCK_ACCOUNT_SMT_LEAF_RECORD: Col = "19";
+/// Column block state record
+pub const COLUMN_BLOCK_STATE_RECORD: Col = "18";
 
 /// chain id
 pub const META_CHAIN_ID_KEY: &[u8] = b"CHAIN_ID";
