+++ crates/store/src/tests/mod.rs
@@ -1,3 +1,3 @@
 mod state_db;
 mod transaction;
-mod transaction_clear_account_state;
+mod transaction_clear_block_state;
